refactor: restucture code
