+++ package.json
@@ -8,11 +8,12 @@
     "test:case": "mocha tests --require hardhat/register --extension ts -g",
     "test:parallel": "mocha --require hardhat/register --recursive --parallel --extension ts --exit",
     "test": "hardhat test",
+    "integration-test": "hardhat test --config hardhat.config.forking.ts",
     "test:feeder": "hardhat test ./tests/AlpacaFeeder.test.ts",
     "test:xAlpaca": "hardhat test ./tests/xALPACA.test.ts",
     "test:grassHouse": "hardhat test ./tests/GrassHouse.test.ts",
     "testnet": "hardhat console --no-compile --network testnet",
-    "mainnetfork:test": "hardhat test --no-compile ./tests/AlpacaFeederWithFairLaunch.test.ts --config hardhat.config.forking.ts",
+    "integration-test:feeder-fairlaunch": "hardhat test --no-compile ./tests/integration/AlpacaFeederWithFairLaunch.test.ts --config hardhat.config.forking.ts",
     "coverage": "hardhat coverage",
     "solhint": "./node_modules/.bin/solhint -f table contracts/**/*.sol",
     "prettier:solidity": "./node_modules/.bin/prettier --write contracts/**/*.sol"
